# utils/swagger/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "ProjectVoice API",
        "description": "REST API для голосового приложения ProjectVoice с поддержкой WebRTC. API предоставляет функционал для управления пользователями, серверами, каналами, сообщениями и приглашениями.",
        "version": "1.0.0",
        "contact": {
            "name": "API Support",
            "email": "support@projectvoice.com"
        }
    },
    "host": "localhost:5000",
    "basePath": "/",
...
    "schemes": ["http", "https"],
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "scheme": "bearer",
            "in": "header",
            "description": "JWT токен для аутентификации. Формат: Bearer {token}"
        }
    },
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/api/invite/{serverId}/invite": {
            "post": {
                "tags": ["Invites"],
                "description": "",
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "expiresAt": {
                                    "example": "any"
                                },
                                "maxUses": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/invite/invite/{token}": {
            "get": {
                "tags": ["Invites"],
                "description": "",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/invite/invite/{token}/accept": {
            "post": {
                "tags": ["Invites"],
                "description": "",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/invite/{serverId}/invites": {
            "get": {
                "tags": ["Invites"],
                "description": "",
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/invite/{inviteId}": {
            "delete": {
                "tags": ["Invites"],
                "description": "",
                "parameters": [
                    {
                        "name": "inviteId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}